Remove redundant dependencies between factorize tasks in H-LU and H-Cholesky
Originally, dependencies were created between the @p factorize task of the current diagonal block and all the @p factorize tasks of its diagonal submatrices. Actually, the dependency on the @p factorize task of its last diagonal submatrix is effective. Hence, the dependencies on other diagonal submatrices are removed. This only remove unnecessary task edges in the flow graph, but does not improve the performance.