Modify prepare_for_vmult_or_tvmult to fall back to serial recursive vmult strategy when the thread num is 1
In some test cases, the problem being solved is very small and there is only one leaf node. Then the thread num computed for task parallel vmult is 1 and the thread data for vmult `data_for_vmult_or_tvmult_threads` will not be initialized, which will cause exception.

Now reset the vmult strategy to serial recursive when the thread num is 1.